rename package to com.deuvox.deuvoxapp
