+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.mvvm_therightway">
+    package="com.deuvox.deuvoxapp">
 
     <application
         android:name=".App"
@@ -9,7 +9,7 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MVVMTheRightWay">
+        android:theme="@style/Theme.Deuvox">
         <activity
             android:name=".feature.main.ExampleMainActivity"
             android:exported="true">
