+++ app/src/main/java/com/deuvox/deuvoxapp/base/anotation/scope/PerActivity.kt
@@ -1,0 +1,7 @@
+package com.deuvox.deuvoxapp.base.anotation.scope
+
+import javax.inject.Scope
+
+@Scope
+@Retention(AnnotationRetention.RUNTIME)
+annotation class PerActivity
