+++ app/src/main/java/com/deuvox/deuvoxapp/base/viewmodel/BaseViewModel.kt
@@ -1,0 +1,14 @@
+package com.deuvox.deuvoxapp.base.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+abstract class BaseViewModel : ViewModel() {
+    override fun onCleared() {
+        super.onCleared()
+        onDispose()
+    }
+
+    open fun onDispose() {
+        // Override to dispose
+    }
+}
