+++ app/src/main/java/com/deuvox/deuvoxapp/base/viewmodel/ViewModelFactory.kt
@@ -1,0 +1,29 @@
+package com.deuvox.deuvoxapp.base.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import javax.inject.Inject
+import javax.inject.Provider
+
+class ViewModelFactory @Inject constructor(
+    private val providers: @JvmSuppressWildcards Map<Class<out ViewModel>, Provider<ViewModel>>
+) : ViewModelProvider.Factory {
+
+    @Suppress("UNCHECKED_CAST")
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        var provider = providers[modelClass]
+        if (provider == null) {
+            for ((key, value) in providers) {
+                if (modelClass.isAssignableFrom(key)) {
+                    provider = value
+                    break
+                }
+            }
+        }
+        if (provider == null) {
+            throw IllegalStateException("unknown model class $modelClass")
+        }
+
+        return provider.get() as T
+    }
+}
