+++ app/src/main/java/com/deuvox/deuvoxapp/di/component/AppComponent.kt
@@ -1,0 +1,11 @@
+package com.deuvox.deuvoxapp.di.component
+
+import com.deuvox.deuvoxapp.App
+import com.deuvox.deuvoxapp.di.module.HttpNetworkModule
+import com.deuvox.deuvoxapp.di.module.ViewModelFactoryModule
+import dagger.Component
+
+@Component(modules = [ViewModelFactoryModule::class, HttpNetworkModule::class])
+interface AppComponent {
+    fun inject(app: App)
+}
