+++ app/src/main/java/com/deuvox/deuvoxapp/di/module/HttpNetworkModule.kt
@@ -1,0 +1,38 @@
+package com.deuvox.deuvoxapp.di.module
+
+import com.google.gson.Gson
+import dagger.Module
+import dagger.Provides
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import java.util.concurrent.TimeUnit
+import javax.inject.Singleton
+
+@Module
+class HttpNetworkModule {
+    @Provides
+    @Singleton
+    fun providesGson(): Gson {
+        return Gson()
+    }
+
+    @Provides
+    @Singleton
+    fun providesHttpClient(): OkHttpClient {
+        return OkHttpClient().newBuilder()
+            .callTimeout(10000, TimeUnit.MILLISECONDS)
+            .readTimeout(10000, TimeUnit.MILLISECONDS)
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun providesRetrofit(httpClient: OkHttpClient): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl("https://example.com/")
+            .addConverterFactory(GsonConverterFactory.create())
+            .client(httpClient)
+            .build()
+    }
+}
