--- app/src/main/java/com/example/mvvm_therightway/base/anotation/scope/PerActivity.kt
@@ -1,7 +1,0 @@
-package com.example.mvvm_therightway.base.anotation.scope
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-annotation class PerActivity
