--- app/src/main/java/com/example/mvvm_therightway/base/anotation/viewmodel/ViewModelKey.kt
@@ -1,15 +1,0 @@
-package com.example.mvvm_therightway.base.anotation.viewmodel
-
-import androidx.lifecycle.ViewModel
-import dagger.MapKey
-import kotlin.reflect.KClass
-
-
-@Target(
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY_GETTER,
-    AnnotationTarget.PROPERTY_SETTER
-)
-@Retention(AnnotationRetention.RUNTIME)
-@MapKey
-internal annotation class ViewModelKey(val value: KClass<out ViewModel>)
