--- app/src/main/java/com/example/mvvm_therightway/di/module/ViewModelFactoryModule.kt
@@ -1,14 +1,0 @@
-package com.example.mvvm_therightway.di.module
-
-import androidx.lifecycle.ViewModelProvider
-import com.example.mvvm_therightway.base.viewmodel.ViewModelFactory
-import dagger.Binds
-import dagger.Module
-import javax.inject.Singleton
-
-@Module
-abstract class ViewModelFactoryModule {
-    @Binds
-    @Singleton
-    abstract fun bindsViewModelFactory(viewModelFactory: ViewModelFactory): ViewModelProvider.Factory
-}
