--- app/src/main/java/com/example/mvvm_therightway/feature/main/ExampleMainViewModel.kt
@@ -1,19 +1,0 @@
-package com.example.mvvm_therightway.feature.main
-
-import android.util.Log
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.domain.GetDataUseCase
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-class ExampleMainViewModel @Inject constructor(
-    private val getDataUseCase: GetDataUseCase
-) : ViewModel() {
-
-    fun getData() {
-        viewModelScope.launch {
-            Log.e("TAGO", "getData: ${getDataUseCase.getData()}")
-        }
-    }
-}
