+++ data/src/main/java/com/deuvox/data/DataNetworkRepository.kt
@@ -1,0 +1,4 @@
+package com.deuvox.data
+
+class DataNetworkRepository {
+}
