--- data/src/main/java/com/example/data/DataNetworkRepository.kt
@@ -1,4 +1,0 @@
-package com.example.data
-
-class DataNetworkRepository {
-}
