+++ domain/src/main/java/com/deuvox/domain/GetDataUseCase.kt
@@ -1,0 +1,7 @@
+package com.deuvox.domain
+
+import javax.inject.Inject
+
+class GetDataUseCase @Inject constructor() {
+    fun getData() = "TestUseCase"
+}
