--- domain/src/main/java/com/example/domain/GetDataUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.domain
-
-import javax.inject.Inject
-
-class GetDataUseCase @Inject constructor() {
-    fun getData() = "TestUseCase"
-}
